+++ apps/router-e2e/app.config.js
@@ -30,8 +30,8 @@ module.exports = {
     typedRoutes: true,
     reactCanary: process.env.E2E_CANARY_ENABLED,
     reactCompiler: process.env.E2E_ROUTER_COMPILER,
-    reactServerComponents: process.env.E2E_RSC_ENABLED,
-    reactServerActions: process.env.EXPO_UNSTABLE_SERVER_ACTIONS,
+    reactServerComponentRoutes: process.env.E2E_RSC_ENABLED,
+    reactServerFunctions: process.env.EXPO_UNSTABLE_SERVER_FUNCTIONS,
   },
   web: {
     output: process.env.EXPO_USE_STATIC ?? 'static',