+++ packages/@expo/cli/e2e/playwright/dev/03-server-actions-only.test.ts
@@ -26,7 +26,7 @@ test.describe(inputDir, () => {
       E2E_ROUTER_JS_ENGINE: 'hermes',
       E2E_ROUTER_SRC: testName,
       E2E_ROUTER_ASYNC: 'development',
-      EXPO_UNSTABLE_SERVER_ACTIONS: '1',
+      EXPO_UNSTABLE_SERVER_FUNCTIONS: '1',
       E2E_CANARY_ENABLED: '1',
       EXPO_USE_METRO_REQUIRE: '1',
       TEST_SECRET_VALUE: 'test-secret',