+++ packages/@expo/cli/CHANGELOG.md
@@ -10,7 +10,11 @@
 
 ### 🐛 Bug fixes
 
+- Fix RSC export with hooks. ([#32793](https://github.com/expo/expo/pull/32793) by [@EvanBacon](https://github.com/EvanBacon))
+
 ### 💡 Others
+
+- Rename `experiments.reactServerActions` -> `experiments.reactServerFunctions` and other RSC flags. ([#32791](https://github.com/expo/expo/pull/32791) by [@EvanBacon](https://github.com/EvanBacon))
 
 ## 0.21.0 — 2024-11-12